changed function definition ret to int

+++ kern/include/file_syscalls.h
@@ -10,8 +10,8 @@
  */
 int open(const_userptr_t filename, int flags, mode_t mode, int *retval);
 int close(int fd);
-ssize_t read(int fd, userptr_t buf, size_t buflen);
-ssize_t write(int fd, const_userptr_t buf, size_t nbytes);
-off_t lseek(int fd, off_t pos, int whence, int *retval, int *retval_v1);
+int read(int fd, userptr_t buf, size_t buflen);
+int write(int fd, const_userptr_t buf, size_t nbytes);
+int lseek(int fd, off_t pos, int whence, int *retval, int *retval_v1);
 
 #endif
